add 104 error handling
